src/requirements.txt
asgiref==3.5.2
Django==4.0.*
djangorestframework==3.13.*
pytz==2022.2
sqlparse==0.4.2
gunicorn==20.1.*
mysqlclient==2.1.*
